rename current devices directory to interfaces
