+++ mio/const.py
@@ -5,4 +5,4 @@ from pathlib import Path
 BASE_DIR = Path(__file__).parent.resolve()
 DATA_DIR = BASE_DIR / "data"
 CONFIG_DIR = DATA_DIR / "config"
-DEVICE_DIR = BASE_DIR / "devices"
+INTERFACES_DIR = BASE_DIR / "interfaces"
