+++ mio/interfaces/XEM7310-A75/USBInterface-3_03mhz-J2_2-3v3-IEEE.bit
