+++ mio/interfaces/XEM7310-A75/USBInterface-6_67mhz-J2_2-3v3-IEEE.bit
