+++ mio/interfaces/__init__.py
@@ -1,0 +1,3 @@
+"""
+Control interfaces for external hardware interfaces
+"""
