+++ mio/stream_daq.py
@@ -18,7 +18,7 @@ from bitstring import BitArray, Bits
 
 from mio import init_logger
 from mio.bit_operation import BufferFormatter
-from mio.devices.mocks import okDevMock
+from mio.interfaces.mocks import okDevMock
 from mio.exceptions import EndOfRecordingException, StreamReadError
 from mio.io import BufferedCSVWriter, VideoWriter
 from mio.models.process import FrequencyMaskingConfig
@@ -38,7 +38,7 @@ BIT_PER_WORD = 32
 okDev = None  # Set if OpalKelly driver is available
 
 try:
-    from mio.devices.opalkelly import okDev
+    from mio.interfaces.opalkelly import okDev
 
     HAVE_OK = True
 except (ImportError, ModuleNotFoundError):
@@ -70,7 +70,7 @@ class StreamDaq:
     --------
     $ mio stream capture -c path/to/config.yml -o output_filename.avi
     Connected to XEM7310-A75
-    Succesfully uploaded /mio/mio/devices/selected_bitfile.bit
+    Succesfully uploaded /mio/mio/interfaces/selected_bitfile.bit
     FrontPanel is supported
 
     .. todo::
@@ -244,13 +244,13 @@ class StreamDaq:
         if read_length is None:
             read_length = int(max(self.buffer_npix) * self.config.pix_depth / 8 / 16) * 16
 
-        # set up fpga devices
+        # set up fpga interfaces
         BIT_FILE = self.config.bitstream
         if not BIT_FILE.exists():
             serial_buffer_queue.put(None)
             raise RuntimeError(f"Configured to use bitfile at {BIT_FILE} but no such file exists")
 
-        # set up fpga devices
+        # set up fpga interfaces
         dev = self._init_okdev(BIT_FILE, read_length)
 
         # read loop
